requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
beautifulsoup4==4.11.1
bs4==0.0.1
cached-property==1.5.2
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
colorama==0.4.4
commonmark==0.9.1
decorator==4.4.2
dictdiffer==0.9.0
future==0.18.2
huepy==1.2.1
idna==3.3
imageio==2.16.1
imageio-ffmpeg==0.4.5
iniconfig==1.1.1
instabot==0.117.0
mock==4.0.3
moviepy==1.0.3
notion==0.0.28
notion-database==20220222.2
numpy==1.22.3
packaging==21.3
pilgram==1.2.1
Pillow==9.1.0
pluggy==1.0.0
proglog==0.1.9
py==1.11.0
pyparsing==3.0.7
PySocks==1.7.1
pytest==7.1.1
python-dotenv==0.20.0
python-slugify==6.1.1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
requests-toolbelt==0.9.1
responses==0.20.0
schedule==1.1.0
six==1.16.0
soupsieve==2.3.2
text-unidecode==1.3
tomli==2.0.1
tqdm==4.64.0
typing-extensions==4.1.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
